--- iq-bech32/CHANGES.md
@@ -1,16 +1,0 @@
-## 0.1.1 (2018-08-29)
-
-* [#74](https://github.com/iqlusioninc/crates/pull/74)
-  Add `Bech32::lower_case()` method.
-
-## 0.1.0 (2018-08-29)
-
-* [#70](https://github.com/iqlusioninc/crates/pull/70)
-  First pass on eliminating secret leakages.
-
-* [#69](https://github.com/iqlusioninc/crates/pull/70)
-  Calculate charset inverses programatically.
-
-## 0.0.1 (2018-08-26)
-
-* Initial release
