--- iq-bech32/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name        = "iq-bech32"
-description = """
-              Hardened Bech32 (BIP-173) implementation: a human-friendly base-32
-              encoding format with checksums to prevent transcription errors.
-              This implementation is hardened for handling Bech32-encoded secret
-              keys by automatically clearing memory containing sensitive data.
-              Coming soon: constant-time encoding/decoding support.
-              """
-version     = "0.1.1" # Also update html_root_url in lib.rs when bumping this
-authors     = ["Tony Arcieri <tony@iqlusion.io>"]
-license     = "Apache-2.0"
-homepage    = "https://github.com/iqlusioninc/crates/"
-repository  = "https://github.com/iqlusioninc/crates/tree/master/iq-bip32"
-readme      = "README.md"
-categories  = ["cryptography", "encoding"]
-keywords    = ["bitcoin", "bech32", "cryptography", "keys", "security"]
-
-[badges]
-circle-ci = { repository = "iqlusioninc/crates" }
-
-[dependencies]
-clear_on_drop = "0.2"
-failure = "0.1"
