+++ subtle-encoding/src/bech32/base32.rs
@@ -1,3 +1,5 @@
+use prelude::*;
+
 use super::Error;
 
 /// Encode binary data as base32
